+++ ChangeLog.md
@@ -8,6 +8,8 @@
              [Issue #115](https://github.com/grobian/carbon-c-relay/issues/115)
 * **server** when connecting to a hostname, now /all/ resolved addresses
              are tried in order before failing
+* **relay** A new `-O` flag allows to set the minimum rules treshold
+            before trying to optimise the rules.
 
 
 # 2.5 (2017-01-09)